Рефакторинг миграций

+++ word-map/src/main/java/com/margot/word_map/repository/map/LetterRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LetterRepository extends JpaRepository<Letter, Short> {
+public interface LetterRepository extends JpaRepository<Letter, Long> {
 }
